requirements.txt
mediapipe
opencv-python
pyaudio
pyyaml